Fixed webpack loading for scrollbar
